Move useAllowScreenshots to utils
