+++ apps/wallet-mobile/src/Catalyst/QrCode.tsx
@@ -5,10 +5,10 @@ import QRCodeSVG from 'react-native-qrcode-svg'
 import {SafeAreaView} from 'react-native-safe-area-context'
 
 import {Button, CopyButton, ProgressStep, Spacer, Text} from '../components'
-import {useAllowScreenshots} from '../features/Settings/ScreenShare'
 import {confirmationMessages} from '../i18n/global-messages'
 import {useBlockGoBack} from '../navigation'
 import {COLORS} from '../theme'
+import {useAllowScreenshots} from '../utils'
 import {Actions, Description, Title} from './components'
 import {useCountdown} from './hooks'
 import {VotingRegistrationBackupCheckModal} from './VotingRegistrationBackupCheckModal'
