+++ apps/wallet-mobile/src/features/Settings/ScreenShare/index.ts
@@ -1,6 +1,1 @@
-export {
-  useAllowScreenshots,
-  useChangeScreenShareSetting,
-  useInitScreenShare,
-  useScreenShareSettingEnabled,
-} from './ScreenShare'
+export {useChangeScreenShareSetting, useInitScreenShare, useScreenShareSettingEnabled} from './ScreenShare'
