+++ apps/wallet-mobile/src/utils/utils.ts
@@ -1,4 +1,11 @@
+import {useFocusEffect} from '@react-navigation/native'
+import {isBoolean} from '@yoroi/common'
 import {isEmpty} from 'lodash'
+import React from 'react'
+import {Platform} from 'react-native'
+
+import {useScreenShareSettingEnabled} from '../features/Settings/ScreenShare'
+import {changeScreenShareNativeSettingOnAndroid} from '../features/Settings/ScreenShare/ScreenShare'
 
 export function isEmptyString(value: string | null | undefined): value is '' | null | undefined {
   return isEmpty(value)
@@ -16,4 +23,18 @@ export function promiseAny<T>(promises: Promise<T>[]): Promise<T> {
       })
     }
   })
+}
+
+export const useAllowScreenshots = () => {
+  const {data: screenShareSettingEnabled} = useScreenShareSettingEnabled()
+  const callback = React.useCallback(() => {
+    if (Platform.OS !== 'android') return
+    if (!isBoolean(screenShareSettingEnabled) || screenShareSettingEnabled) return
+
+    changeScreenShareNativeSettingOnAndroid(true)
+    return () => {
+      changeScreenShareNativeSettingOnAndroid(false)
+    }
+  }, [screenShareSettingEnabled])
+  useFocusEffect(callback)
 }
