+++ apps/wallet-mobile/translations/messages/src/Catalyst/QrCode.json
@@ -6,12 +6,12 @@
     "start": {
       "line": 92,
       "column": 12,
-      "index": 2914
+      "index": 2890
     },
     "end": {
       "line": 95,
       "column": 3,
-      "index": 3012
+      "index": 2988
     }
   },
   {
@@ -21,12 +21,12 @@
     "start": {
       "line": 96,
       "column": 15,
-      "index": 3029
+      "index": 3005
     },
     "end": {
       "line": 99,
       "column": 3,
-      "index": 3151
+      "index": 3127
     }
   },
   {
@@ -36,12 +36,12 @@
     "start": {
       "line": 100,
       "column": 16,
-      "index": 3169
+      "index": 3145
     },
     "end": {
       "line": 105,
       "column": 3,
-      "index": 3399
+      "index": 3375
     }
   },
   {
@@ -51,12 +51,12 @@
     "start": {
       "line": 106,
       "column": 16,
-      "index": 3417
+      "index": 3393
     },
     "end": {
       "line": 111,
       "column": 3,
-      "index": 3635
+      "index": 3611
     }
   },
   {
@@ -66,12 +66,12 @@
     "start": {
       "line": 112,
       "column": 8,
-      "index": 3645
+      "index": 3621
     },
     "end": {
       "line": 115,
       "column": 3,
-      "index": 3793
+      "index": 3769
     }
   },
   {
@@ -81,12 +81,12 @@
     "start": {
       "line": 116,
       "column": 14,
-      "index": 3809
+      "index": 3785
     },
     "end": {
       "line": 119,
       "column": 3,
-      "index": 3900
+      "index": 3876
     }
   }
 ]
